fix rubocop violations
