+++ enterprise/app/helpers/captain/chat_helper.rb
@@ -1,5 +1,6 @@
 module Captain::ChatHelper
   include Integrations::LlmInstrumentation
+  include Captain::ChatResponseHelper
 
   def request_chat_completion
     log_chat_completion_request
@@ -87,55 +88,6 @@ module Captain::ChatHelper
 
   def conversation_messages
     @messages.reject { |m| m[:role] == 'system' || m[:role] == :system }
-  end
-
-  def persist_thinking_message(tool_call)
-    return unless defined?(@copilot_thread) && @copilot_thread.present?
-
-    tool_name = tool_call.name.to_s
-
-    persist_message(
-      {
-        'content' => "Using #{tool_name}",
-        'function_name' => tool_name
-      },
-      'assistant_thinking'
-    )
-  end
-
-  def persist_tool_completion
-    return unless defined?(@copilot_thread) && @copilot_thread.present?
-
-    tool_call = @pending_tool_calls&.pop
-    return unless tool_call
-
-    tool_name = tool_call.name.to_s
-
-    persist_message(
-      {
-        'content' => "Completed #{tool_name}",
-        'function_name' => tool_name
-      },
-      'assistant_thinking'
-    )
-  end
-
-  def build_response(response)
-    Rails.logger.debug { "#{self.class.name} Assistant: #{@assistant.id}, Received response #{response}" }
-
-    parsed = parse_json_response(response.content)
-
-    persist_message(parsed, 'assistant')
-    { 'response' => parsed['content'] }
-  end
-
-  def parse_json_response(content)
-    content = content.gsub('```json', '').gsub('```', '')
-    content = content.strip
-    JSON.parse(content)
-  rescue JSON::ParserError => e
-    Rails.logger.error "#{self.class.name} Assistant: #{@assistant.id}, Error parsing JSON response: #{e.message}"
-    { 'content' => content }
   end
 
   def temperature
